Start to use arpeggio.
