+++ KustoPandas/expression_parser/__init__.py
@@ -1,3 +1,7 @@
-from .expression_parser import parse_expression
+from .arpeggio_parser import parse_expression, parse_expression_toplevel
 from .expression_parser_types import Assignment, Method, Var, By, Comma, Star, Mul, Asc, Desc
 from .expression_tree import flatten_comma
+
+from ._simple_expression import SimpleExpression, _get_method_default_name, _generate_temp_column_name, replace_temp_column_names
+
+from .tabular_operators import Take, Extend, Where
